moved startpage to its own comp
